+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+  - `&TextDelta.lines/1` and `&TextDelta.lines!/1`
 
 ## [1.2.0] - 2017-05-29
 ### Added